requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pandas==2.2.3
python-multipart==0.0.20
pydantic==2.10.6
pydantic-settings==2.8.1
pyjwt==2.10.1
pytest==8.3.5
kagglehub==0.3.10
python-dotenv==1.0.1
sqlalchemy==2.0.39
passlib==1.7.4
bcrypt==4.0.1
tf-keras==2.16.0
transformers==4.46.2
datasets==3.3.2
tensorflow==2.16.1
tensorboard==2.16.0
pytest-asyncio==0.25.3
httpx==0.28.1